Move list into new component
