+++ src/App.jsx
@@ -2,32 +2,18 @@ import { useState } from 'react';
 import reactLogo from './assets/react.svg';
 import viteLogo from '/vite.svg';
 import './App.css';
+import TodoList from './TodoList';
 
 function App() {
   // const [count, setCount] = useState(0)
-  let todoList = [
-    {
-      id: 1,
-      title: 'Complete Assignment',
-    },
-    {
-      id: 2,
-      title: 'Read Supplemental Materials',
-    },
-    {
-      id: 3,
-      title: 'Check Slack',
-    },
-  ];
+
 
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(({ id, title }) => (
-          <li key={id}>{title}</li>
-        ))}
-      </ul>
+
+      <TodoList />
+
     </div>
   );
 }
